Create new sub folder to kubernets test
